requirements.txt
# Dataframe packages
pandas
numpy

# Data visualization packages
matplotlib
seaborn
plotly

# Frontend packages
streamlit

# Video processing
opencv-python

# SQL packages
mysql-connector-python
pymysql
SQLAlchemy

dotenv

# Yolo
ultralytics

# Rag
langchain
langchain_community
langchain-ollama
ollama

# Google Gen Ai
google-generativeai